Added page 500 to RootBoundary. Corrected a typo.
