+++ packages/client/src/services/y-api-service.ts
@@ -1,9 +1,13 @@
-import { RegistrationFromApi, UserFromApi } from '../types/FromApi'
+import {
+  RegistrationFromApi,
+  UserFromApi,
+  LoginFromApi,
+} from '../types/FormApi'
 
 export const Y_API_BASE_URL = 'https://ya-praktikum.tech/api/v2'
 
 const yApiService = {
-  login(userData: RegistrationFromApi): Promise<Response> {
+  login(userData: LoginFromApi): Promise<Response> {
     return fetch(`${Y_API_BASE_URL}/auth/signin`, {
       method: 'POST',
       headers: {
